Add .gitattributes to control line endings in checked out files
XML processors normalize line breaks, and thus convert the combination
of carriage return and line feed to a line feed, see also
https://www.w3.org/TR/xml/#sec-line-ends. The index.html files in the
top-level directory, in the subregister directories and in the code
list directories are regenerated every time
generate-code-list-register-site.bat (in genericode-transformer) is
run. Even when the actual, textual contents haven't changed, `git diff`
will indicate that the files are changed if the automatic settings for
line endings are used, because the index.html will have CRLF line
endings when checked out from git, but the regenerated index.html will
have LF line endings after the XSLT and XProc processors have worked on
it.